[shared api] Extract provisionserver funcs into common_provisionserver
